refactor: more type imports
